feat(HTD-94): move testimonials and statistics to common sections
