+++ src/app/(alternatives)/layout.tsx
@@ -2,11 +2,11 @@ import { Suspense } from 'react';
 
 import { MockResponseFeatures } from '@/components/common-sections/mock-response-features';
 import { RewriteAnything } from '@/components/common-sections/rewrite-anything';
+import { Statistics } from '@/components/common-sections/statistics';
+import { Testimonials } from '@/components/common-sections/testimonials';
 import { TryItForYourselfCTA } from '@/components/common-sections/try-it-for-yourself';
 import { Layout } from '@/components/layout';
 import { HttpToolkitFeatures } from '@/components/sections/alternatives/http-toolkit-features';
-import { Statistics } from '@/components/sections/statistics';
-import { Testimonials } from '@/components/sections/testimonials';
 
 export default function AlternativesLayout({ children }: { children: React.ReactNode }) {
   return (
@@ -20,7 +20,7 @@ export default function AlternativesLayout({ children }: { children: React.React
         <Testimonials />
       </Suspense>
       <Suspense>
-        <Statistics title="Why `*HTTP Toolkit*`?" text="Numbers that speak for themselves:" />
+        <Statistics />
       </Suspense>
       <TryItForYourselfCTA variant="cta-fluid" />
     </Layout>
