+++ src/components/common-sections/statistics/stadistics.types.ts
@@ -1,4 +1,4 @@
 export interface StatisticsProps {
-  title: string;
-  text: string;
+  title?: string;
+  text?: string;
 }
